slightly refactor workspace tests
